Layout the Moho
Along the Moho, we accept initially-parsed syntax in the form of
raw application expressions and declarations and re-parse them
into forms more amenable to semantic analysis, resolving and binding
names along the way.  The final target of the phase is a much simpler,
more reasonable AST that should point back into the original for
the final phase that emits user-relevant diagnostics: TypeChecking.