First pass at the type checker
Initial import of the sources needed to define the elaboration, type
checking, and solving phases that define the operations of the Mantle.
Broadly, this patch accomplishes three goals

- A Type Checking Phase

The phase responsible for the dictating the structure of lowering user
code to the internal type theory.  Also acts as the place where the
results of the other phases can be sanity-checked, and where most
diagnostics should be produced (for the time being).

- An Elaborator

This phase implements the middle-end lowering of syntax to TT.  Along
the way it generates heterogeneous constraints that can later be broken
into homogeneous constraints and solved by Pattern Unification.

- A Solver

This phase implements the back-end elimination of metavariables
introduced by the elaborator by Gundry-McBride Dynamic Pattern
Unification.  It also supports a number of term reduction techniques
outlined by Pientka and is written in a modular way to support more in
the future.

The intent is not to have a perfect type checker yet, merely to stand
something up that is capable of checking the most basic declarations and
that gives us room to grow in the future.