Provide an explicit iterator to Syntax Collection
The default IndexingIterator for a custom collection is not suitable for
subclasses that specialize the element type
the way libSyntax does.  Overriding it with a more specific iterator,
even with something like covariant overrides, would not work because it
is generic over the collection, not its element type.

Provide a custom iterator parameterized by the element type instead.