Properly implement reparsing
Follow Danielsson and Norell's example by completing the parser for
mixfix operators.  A mixfix operator is composed of sections that are
either wild (_) or identifiers.  With these and a set of ordered
precedences, we can break an expression down into its base tokens and
"reparse" it into a form that Scope Check can better understand.

More work needs to be done on diagnostics and recovery for this phase.
Right now the reparser stops when it encounters an ambiguity.  This
means it preserves the index at which a future diagnostic pass should
start at to determine what went wrong.