Add EmptyDataDeclSyntax and remove '|' (#74)
* Add EmptyDataDeclSyntax and remove '|'

This brings the syntax more in line with Agda, and splitting empty data
decls gives us more flexibility when doing parser recovery to detect bad
constructors.

The strategy here: If you have a data declaration with a `where`, you
best have a constructor next to it. If not, that's an error. If there's
a function directly after one such data declaration, suggest indenting
it to make it a constructor.

* Fix grammar

* Simplify grabbing where token