Implement an indirect return convention
Compile some simple cases of polymorphism with it

The idea is that CPS makes it really messy to escape a by-address copy
because the lifetime of the buffer would have to transfer to the
continuation which may be in the calling frame.  This means we can’t
alloca - so we must malloc everywhere.

Instead, bless a special parameter at the end of the function for
address-only returns.  This buffer is allocated locally by the caller,
assumed alive in the callee, and destroyed by the return continuation.