Add high-level abstractions for manipulating address values
Add the semantic Address types and the Explosion type

An Address value represents an IRValue that is known to be an address value.  The underlying pointer comes paired with an alignment value that is helps keep us honest during codegen.

Address is specialized to OwnedAddress, for heap allocations managed by the Silt runtime, and StackAddress, for managed stack allocations.