Add high level abstractions for manipulating value groups
Steal the concept of Explosion wholesale from Swift.

An Explosion is a construct that closess over a number of values, and only allows for the linear consumption of those values by operations or by other explosions.  It is a fatal condition for an explosion to pass out of scope without being consumed.

To help manage the form of an explosion, and to gain information about them without fully realizing them, also add an Explosion Schema type.