Lay the ground work for "Value Cohabitation"
When laying a datatype out in memory, very often there are bitpatterns that do not correspond prima fascie to valid values.  For example, due to alignment restrictions on x86-64, the bottom 3 bits of any pointer value are guaranteed to be zero.  In fact, only 56 bits in total are in use despite pointer values being ostensibly "64 bits wide".  We can use these "spare bits" to compress the structure of data types.  Because we are using the invalid bitpatterns of one type to compress another, usually the outer type, we refer to these values as "cohabitations".

For example, a single-payload datatype leaves us spare bits available in which we can place the tag bits.

Add the Cohabitable protocol to and begin adding spare bits, spare bitmasks, and cohabitant counts to the data type strategies.