Stub out an optimizer
The idea is to have "Scope passes" that run late on scheduled GIR, and some kind of "Sea of Node passes" that run on the unscheduled GIR.  Quite what the latter looks like, I have no idea.

In the mean time, stub out the pass pipeliner, and silt-optimize tool so we can test this.  Add the beginnings of a SimplifyCFG pass and teach it how to remove dead blocks.  We definitely want to implement Jump Threading ASAP.