XFAIL the Linux build
We're currently caught between a rock and a hard place: The Swift 5.0
compiler crashes building the type checker on Linux.  So we upgrade to
Swift 5.1 snapshots.  But those drop include flags from dependent
targets (see https://bugs.swift.org/browse/SR-10576) so we're SOL.

Disable the Linux build to keep things moving on