Fix pattern specialization
Fixing the scoping issues revealed an unsurprising bug: We're creating dummy pattern variables for constructor patterns that have their de-bruijn indices ass-backwards.  This doesn't affect type checking because we don't look at the clause row for that.  But it does make GIRGen a lot more brittle.

I'm not sure what all to do about this because we need to examine patterns forwards, and those patterns might contain subpatterns that bind additional variables into scope.  For now, we'll just live with it and work around it when binding pattern variables in GIRGen.

This is less than ideal.